Revert "Support matching of PASS resources by IP address/port pairs (#83)"
This reverts commit 83b06f5079a665661c6759fb3f2d76163f33346c.

Elasticsearch query string escaping is complex.  This PR resulted in
searches involving multiple terms (with spaces) to fail.  As a result,
the bet course of action is to keep ES queries via the existing client
methods as literal, quoted strings and use other methods (including
manually performing ES queries, or adding a new API method in the
client)